jito_protos

Module block_engine

Modules

block_engine_relayer_client

Generated client implementations.

block_engine_relayer_server

Generated server implementations.

block_engine_validator_client

Generated client implementations.

block_engine_validator_server

Generated server implementations.

packet_batch_update

Nested message and enum types in PacketBatchUpdate.

Structs

AccountsOfInterest

AccountsOfInterestRequest

AccountsOfInterestUpdate

BlockBuilderFeeInfoRequest

BlockBuilderFeeInfoResponse

ExpiringPacketBatch

A series of packets with an expiration attached to them. The header contains a timestamp for when this packet was generated. The expiry is how long the packet batches have before they expire and are forwarded to the validator. This provides a more censorship resistant method to MEV than block engines receiving packets directly.

PacketBatchUpdate

Packets and heartbeats are sent over the same stream. ExpiringPacketBatches have an expiration attached to them so the block engine can track how long it has until the relayer forwards the packets to the validator. Heartbeats contain a timestamp from the system and is used as a simple and naive time-sync mechanism so the block engine has some idea on how far their clocks are apart.

ProgramsOfInterestRequest

ProgramsOfInterestUpdate

StartExpiringPacketStreamResponse

SubscribeBundlesRequest

SubscribeBundlesResponse

SubscribePacketsRequest

SubscribePacketsResponse