jito_protos::block_engine::block_engine_validator_server

Trait BlockEngineValidator

pub trait BlockEngineValidator:
    Send
    + Sync
    + 'static {
    type SubscribePacketsStream: Stream<Item = Result<SubscribePacketsResponse, Status>> + Send + 'static;
    type SubscribeBundlesStream: Stream<Item = Result<SubscribeBundlesResponse, Status>> + Send + 'static;

    // Required methods
    fn subscribe_packets<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscribePacketsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribePacketsStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_bundles<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscribeBundlesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeBundlesStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_builder_fee_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlockBuilderFeeInfoRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BlockBuilderFeeInfoResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with BlockEngineValidatorServer.

Required Associated Types

type SubscribePacketsStream: Stream<Item = Result<SubscribePacketsResponse, Status>> + Send + 'static

Server streaming response type for the SubscribePackets method.

type SubscribeBundlesStream: Stream<Item = Result<SubscribeBundlesResponse, Status>> + Send + 'static

Server streaming response type for the SubscribeBundles method.

Required Methods

fn subscribe_packets<'life0, 'async_trait>( &'life0 self, request: Request<SubscribePacketsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribePacketsStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/ Validators can subscribe to the block engine to receive a stream of packets

fn subscribe_bundles<'life0, 'async_trait>( &'life0 self, request: Request<SubscribeBundlesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeBundlesStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/ Validators can subscribe to the block engine to receive a stream of simulated and profitable bundles

fn get_block_builder_fee_info<'life0, 'async_trait>( &'life0 self, request: Request<BlockBuilderFeeInfoRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BlockBuilderFeeInfoResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Block builders can optionally collect fees. This returns fee information if a block builder wants to collect one.