jito_protos

Module bundle

Modules

bundle_result

Nested message and enum types in BundleResult.

rejected

Nested message and enum types in Rejected.

Structs

Accepted

Indicates the bundle was accepted and forwarded to a validator. NOTE: A single bundle may have multiple events emitted if forwarded to many validators.

Bundle

BundleResult

BundleUuid

Dropped

DroppedBundle

Bundle dropped (e.g. because no leader upcoming)

Finalized

InternalError

Bundle dropped due to an internal error.

Processed

Rejected

Indicates the bundle was dropped and therefore not forwarded to any validator.

SimulationFailure

Bundle dropped due to simulation failure.

StateAuctionBidRejected

Indicates the bundle’s bid was not high enough to be included in its state auction’s set of winners.

WinningBatchBidRejected

Indicates the bundle’s bid was high enough to win its state auction. However, not high enough relative to other state auction winners and therefore excluded from being forwarded.

Enums

DroppedReason