jito_protos::searcher::searcher_service_client

Struct SearcherServiceClient

pub struct SearcherServiceClient<T> { /* private fields */ }

Implementations

impl SearcherServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> SearcherServiceClient<T>
where T: GrpcService<Body>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> SearcherServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<Body>, Response = Response<<T as GrpcService<Body>>::ResponseBody>>, <T as Service<Request<Body>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn subscribe_bundle_results( &mut self, request: impl IntoRequest<SubscribeBundleResultsRequest>, ) -> Result<Response<Streaming<BundleResult>>, Status>

Searchers can invoke this endpoint to subscribe to their respective bundle results. A success result would indicate the bundle won its state auction and was submitted to the validator.

pub async fn send_bundle( &mut self, request: impl IntoRequest<SendBundleRequest>, ) -> Result<Response<SendBundleResponse>, Status>

pub async fn get_next_scheduled_leader( &mut self, request: impl IntoRequest<NextScheduledLeaderRequest>, ) -> Result<Response<NextScheduledLeaderResponse>, Status>

Returns the next scheduled leader connected to the block engine.

pub async fn get_connected_leaders( &mut self, request: impl IntoRequest<ConnectedLeadersRequest>, ) -> Result<Response<ConnectedLeadersResponse>, Status>

Returns leader slots for connected jito validators during the current epoch. Only returns data for this region.

pub async fn get_connected_leaders_regioned( &mut self, request: impl IntoRequest<ConnectedLeadersRegionedRequest>, ) -> Result<Response<ConnectedLeadersRegionedResponse>, Status>

Returns leader slots for connected jito validators during the current epoch.

pub async fn get_tip_accounts( &mut self, request: impl IntoRequest<GetTipAccountsRequest>, ) -> Result<Response<GetTipAccountsResponse>, Status>

Returns the tip accounts searchers shall transfer funds to for the leader to claim.

pub async fn get_regions( &mut self, request: impl IntoRequest<GetRegionsRequest>, ) -> Result<Response<GetRegionsResponse>, Status>

Returns region the client is directly connected to, along with all available regions